Reorganized file structure.
